feat(view): add grammar input view (EBNFInputView) for RecursiveAutomaton creation, update messages.properties
