+++ src/main/resources/messages.properties
@@ -254,3 +254,8 @@ RecursiveAutomaton=recursive automaton
 RecursiveAutomaton.Untitled=Untitled
 RecursiveAutomaton.Deterministic=deterministic
 RecursiveAutomaton.Nondeterministic=nondeterministic
+RecursiveAutomatonFactory.FromCFG=From context-free grammar
+RecursiveAutomatonFactory.FromCFGButton=Enter CFG
+EBNF.Input.Title=CFG in EBNF format
+EBNF.Input.Info=Enter context-free grammar in EBNF format
+EBNF.Input.Error=Make sure you leave no blank fields and add at least one production.
